News: Python class created, endpoints added
This commit introduces the news object in openSchool - a way for administrators to tell information about their organization to the users without having to create blogs or set up newsletters. Along with the Python class used for internal representation, there are also new private API endpoints for manipulating news contents without digging in files and modifying the database manually.

Currently, news objects are not stored in a database, but in files. They can be found in the '''cdn/''' folder. But we are planning on moving news storage into the database, which would simplify creation, manipulation, storage and deletion.

We are also working on the documentation for the new endpoints, but due to the ongoing translation work by our (lazy) translator, that would take some time.