`Notes` class deprecated, automatic folder creation
The `Notes` class code is removed from the `classes` module. The reason for deprecation is that the code was a functional mirror of the `Homework` class, the endpoints were missing and there was no use for that code.

Now authentication folders (i.e. `codes/`, `tokens/`) are created automatically upon the import of the `classes` module.