Patch: CDN endpoint not satitizing encoded paths
This is a hotfix for a vulnerability that makes use of the `cdn/` endpoint. If the user-provided path is encoded, the built-in Flask handlers don't follow the path. As a result, a user can access files that are above the `cdn/` folder on the system. Since Core is also run under root, system files could be also accessed.

Now the CDN endpoint encodes slashes and follows the directory, following the behaviour of Flask path handlers. In case the path tries to access something about the CDN folder, the endpoint returns error 404.